jQuery Scala Stack - Refactor - New Scala JS lib adopted for Jetty logging - Clean up
