+++ pom.xml
@@ -40,8 +40,8 @@
       <version>3.1.0</version>
     </dependency>
     <dependency>
-      <groupId>io.cenet.compiler.servlet</groupId>
-      <artifactId>servlet-scalajs-compiler</artifactId>
+      <groupId>com.github.aimmoth.compiler.jetty</groupId>
+      <artifactId>jetty-scalajs-compiler</artifactId>
       <version>1.0-SNAPSHOT</version>
     </dependency>
     <dependency>
@@ -51,8 +51,8 @@
     </dependency>
     <dependency>
       <groupId>org.glassfish.jersey.containers</groupId>
-      <!-- if your container implements Servlet API older than 3.0, use "jersey-container-servlet-core" -->
       <artifactId>jersey-container-servlet</artifactId>
+      <!-- if your container implements Servlet API older than 3.0, use "jersey-container-servlet-core" -->
       <!-- <artifactId>jersey-container-servlet-core</artifactId> -->
       <version>2.18</version>
     </dependency>
