+++ src/main/webapp/WEB-INF/logging.properties
@@ -10,4 +10,7 @@
 #
 
 # Set the default logging level for all loggers to WARNING
-example.level = FINE
+api.level = FINE
+datastore.level = FINE
+servlet.level = FINE
+shared.level = FINE
