+++ src/main/webapp/scalajs/shared/Resource.scala
@@ -1,4 +1,4 @@
-package script
+package shared
 
 class Resource {
