+++ src/main/webapp/scalajs/shared/ResourceValidator.scala
@@ -1,4 +1,4 @@
-package script
+package shared
 
 object ResourceValidator {
